Fix gemspec
It wasn't possible to bundle install the dependencies, as this command
tries to load the gemspec, the gemspec required the code, and the code
was trying to load the dependencies that were not yet available. This
fix is to just load the file that defines the version number from the
gemspec, not the whole code.